+++ CHANGELOG.md
@@ -31,6 +31,7 @@ All notable changes to this project will be documented in this file.
 - Bump goodcheck from 2.4.4 to 2.4.5 [#591](https://github.com/sider/runners/pull/591)
 - Hide userinfo of HTTP URI from the traces [#585](https://github.com/sider/runners/pull/585)
 - [PHPCS] Empower runner [#592](https://github.com/sider/runners/pull/592)
+- Use Yarn when duplicate lock files are found [#593](https://github.com/sider/runners/pull/593)
 
 ## 0.12.0